+++ src/collective/excelexport/testing.py
@@ -1,75 +1,74 @@
 # -*- coding: utf-8 -*-
 """Base module for unittesting."""
 
-import unittest
 
-from plone.app.testing import applyProfile
+from plone import api
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import login
-from plone.app.testing import PLONE_FIXTURE
-from plone.app.testing import PloneSandboxLayer
-from plone.app.testing import setRoles
+from plone.app.testing import PloneWithPackageLayer
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import TEST_USER_NAME
-from plone.testing import z2
+from plone.app.testing import login
+from plone.app.testing import setRoles
 
 import collective.excelexport
+import pkg_resources
+import unittest
+
+try:
+    pkg_resources.get_distribution('plone.app.contenttypes')
+except pkg_resources.DistributionNotFound:
+    HAS_PA_CONTENTTYPES = False
+else:
+    HAS_PA_CONTENTTYPES = True
 
 
-class CollectiveExcelexportLayer(PloneSandboxLayer):
+class CollectiveExcelexportLayer(PloneWithPackageLayer):
 
-    defaultBases = (PLONE_FIXTURE,)
-    products = ('collective.excelexport',
-               )
-
-    def setUpZope(self, app, configurationContext):
-        """Set up Zope."""
-        # Load ZCML
-        self.loadZCML(package=collective.excelexport,
-                      name='testing.zcml')
-        import plone.app.dexterity
-        self.loadZCML(package=plone.app.dexterity,
-                      name='configure.zcml')
-        for p in self.products:
-            z2.installProduct(app, p)
+    def setUpZope(self, *args, **kwargs):
+        """Prepare Zope instance"""
+        super(CollectiveExcelexportLayer, self).setUpZope(*args, **kwargs)
+        if HAS_PA_CONTENTTYPES:
+            import plone.app.contenttypes
+            self.loadZCML(package=plone.app.contenttypes)
 
     def setUpPloneSite(self, portal):
         """Set up Plone."""
-        # Install into Plone site using portal_setup
-        applyProfile(portal, 'collective.excelexport:testing')
+        super(CollectiveExcelexportLayer, self).setUpPloneSite(portal)
+
+        # Plone 5 support
+        if HAS_PA_CONTENTTYPES:
+            self.applyProfile(portal, 'plone.app.contenttypes:default')
 
         # Login and create some test content
         setRoles(portal, TEST_USER_ID, ['Manager'])
         login(portal, TEST_USER_NAME)
-        folder_id = portal.invokeFactory('Folder', 'folder')
-        portal[folder_id].reindexObject()
 
-        # Commit so that the test browser sees these objects
-        import transaction
-        transaction.commit()
-
-    def tearDownZope(self, app):
-        """Tear down Zope."""
-        for p in reversed(self.products):
-            z2.uninstallProduct(app, p)
+        api.content.create(
+            container=portal,
+            type='Folder',
+            id='folder',
+        )
 
 
 FIXTURE = CollectiveExcelexportLayer(
-    name="FIXTURE"
-    )
+    zcml_package=collective.excelexport,
+    zcml_filename='testing.zcml',
+    gs_profile_id='collective.excelexport:testing',
+    name="FIXTURE",
+)
 
 
 INTEGRATION = IntegrationTesting(
-    bases=(FIXTURE,),
+    bases=(FIXTURE, ),
     name="INTEGRATION"
-    )
+)
 
 
 FUNCTIONAL = FunctionalTesting(
-    bases=(FIXTURE,),
+    bases=(FIXTURE, ),
     name="FUNCTIONAL"
-    )
+)
 
 
 class IntegrationTestCase(unittest.TestCase):